election-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
jinja2==3.1.2
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
